requirements.txt
streamlit>=1.32.0
Pillow>=10.2.0
qrcode[pil]>=7.4.2
reportlab>=4.1.0
numpy>=1.26.0
requests>=2.31.0
fonttools>=4.50.0
